Julie Corrected a typing mistake in the call to XERBLA
Julie: Modifications so the tests runs fine.

xerrgg.f : XGGESX has been previously modified to allow a workspace query for IWORK with LIWORK = -1, but the tester XERRGG expects that to be an error (line 641 of xerrgg.f).  Set argument 24 to 0 instead of -1 in that test

xerrst.f : XDSYEVX has been previously modified to improve the calculation of the workspace. When N = 1, LWORK now only needs to be 1, instead of 8.
set LWORK = 0, to be safe.