Jason's Bugs with Michael Pont fix ----------------------------------
REMOVE Jason's correction [clatm1.f and zlatm1.f]

SLARAN:
Bug description:
the bug is actually in function SLARAN, which can sometimes
return exactly 1.0 (it says it won't). The exact 1.0 causes
CLARND to return exactly (0.0,0.0) when idist=3.
N.B. CLARND can legitimately return exactly (0.0,0.0) if
called with idist=2 or idist=4, but CLATM1 only calls with
idist=3.

If a real number has n bits of precision, and the first
n bits of the 48-bit integer above happen to be all 1 (which
will occur about once every 2**n calls), then SLARAN will
be rounded to exactly 1.0. In IEEE single precision arithmetic,
this will happen relatively often since n = 24.
Since SLARAN is not supposed to return exactly 0.0 or 1.0
(and some callers of SLARAN, such as CLARND, depend on that),
the statistically correct thing to do in this situation is
simply to iterate again.
N.B. the case SLARAN = 0.0 should not be possible.

apply the same fix for dlaran