Emphasize SLACN2 over SLACON.
Convert the examples in Chapter 4 to use the new, thread-
safe SLACN2 rather than SLACON.  Adds {S,C}LACN2 to the
Index of Auxilliary Routines in lug.tex, explaining that
the old ones are not thread-safe.  Also update the mention
in the "manpage" for STGSYL/CTGSYL.

The manpages need regenerated and possibly rotated in such
a way that at least PDF gets it right.

Signed-off-by: Jason Riedy <ejr@cs.berkeley.edu>