Modified files:       SRC/dgehrd.f       SRC/zlahrd.f       SRC/sgehrd.f       SRC/zgehrd.f       SRC/dlahrd.f       SRC/slahrd.f
Add the new Hessenberg from Daniel Kresner (the van de Geijn's variant)

Comments from Daniel:
A pleasant feature of the code is that only very few changes were
necessary; the interface and the amount of required workspace are
exactly the same. An unpleasant feature is that the output of the
modified DLAHRD slightly differs from the original
DLAHRD, i.e., this routine is NOT backward compatible.