Avoid using undefined memory in xSYEVR, xHEEVR, xSTEVR; document indices
SYEVR, HEEVR, and STEVR do not use IWORK if STEGR/STERF succeed,
and the drivers do not return the indices of the vectors that
failed to converge.  However, the sorting loop at the end reorders
the possibly undefined indicators.  IWORK is not documented to
include the failing indices, so we simply remove IWORK from the
sorting loops.

We also initialize and document all the symbolic offsets.