Convert SYEVR, STEVR, HEEVR to use STEMR rather than STEGR-compat.
If possible, we should use only new interfaces and not
compatibility interfaces.

TRYRAC is set to .TRUE. if ABSTOL is at most 2*N*eps.
That choice is based on the recommendation in the drivers'
comments of ABSTOL=eps for accuracy and also the bounds in
STEMR.  The extra factor of 2 is just an additional fudge
factor for my paranoia.

Signed-off-by: Jason Riedy <ejr@cs.berkeley.edu>