LANEG: Use the Inf/Inf=>1 substitution from LAWN 172.
LANEG now imposes no input range restrictions, but other uses
of PIVMIN in the code still require the current restriction
that the input values are less than sqrt(overflow) in magnitude.

The original code differs from the base of LAWN 172's code.  The
code here had an additional test in each "safe" loop,
     IF( TMP.EQ.ZERO ) S = LLD( J )
where TMP = LLD(j) / DPLUS.  I cannot see what it accomplishes.
There are no comments explaining these tests, either.

Both Prof. Kahan and I convinced ourselves that we need the division
to be TMP = T / DPLUS and not TMP = LLD(J) / DPLUS to show
monotonicity, but neither of us has followed through with a complete
proof.  The code is reorganized to support the conjectures.

The twist index used the variable S, which is kept as T + SIGMA.  To
simplify the loops, S is removed and the twist index computation is
shifted back by adding SIGMA.

A good compiler should be limited mostly by data dependencies, so we
can place the test for the pivot's sign immediately after the pivot
itself.

This is not blocked, so machines with very slow Inf/NaN arithmetic
(e.g. Pentium 4s) will see horrible performance if exceptions occur.

Signed-off-by: Jason Riedy <ejr@cs.berkeley.edu>