Correction submitted by Sven and Daniel.
we agree that this is probably what is intended for these routines, and it avoids the undefined references to SSLECT and DSLECT.

Here is a note from Daniel
--------------------------
SSLECT is supposed to return whether a complex number zR + i*zI 
corresponds to a selected eigenvalue. If SELOPT = 0, this is simply done 
by checking whether zR < 0 (-> invariant subspace for stable eigenvalues). 
If SELOPT > 0, I think it wants to find the eigenvalue closest to zR + 
i*zI and checks whether this eigenvalue is selected (by looking up the 
common array SELVAL). However, the following lines

         RMIN = ZERO
         DO 10 I = 1, SELDIM
            X = SLAPY2( ZR-SELWR( I ), ZI-SELWI( I ) )
            IF( X.LE.RMIN ) THEN 

make little sense to me. RMIN should initially be set to a large number or 
SLAPY2( ZR-SELWR( 1 ), ZI-SELWI( 1 ) ). Otherwise, the test will certainly 
fail for real matrices, because the recomputed reordered eigenvalue will 
not correspond exactly to an original eigenvalue if there are 2x2 blocks 
in the Schur form.

The common arrays SELWR, SELWI, SELVAL control the selection of 
eigenvalues. They are set in SGET24 such that NSEL eigenvalues having 
smallest real part are selected. SSGET24 is used in SDRVSX but not in 
SDRVES (which uses SELOPT = 0).

> > Surprisingly, the complex version does not fail with CSLECT undefined.

The reordered eigenvalues for complex matrices correspond exactly to the 
original eigenvalues (see CTREXC).