Initialize the variable USEDQD in xLARRE before use. (Cf. rev430 in the trunk)
* Bug report from Michel Devel from the UTINAM institute, University of
  Franche-Comte, UMR CNRS 6213.
* Patch from Christof and Julien.
* See forum discussion https://icl.cs.utk.edu/lapack-forum/viewtopic.php?t=350

Michel Devel from the UTINAM institute, University of Franche-Comte, UMR CNRS
6213 reports that the variable USEDQD is used while not initialized in xLARRE.
This bug has been found using the Intel Fortran compiler with runtime checks on
while executing the LAPACK testers.

The patch is to add at the beginning of the code, the initialization:

*     Initialize USEDQD, DQDS should be used for ALLRNG unless someone
*     explicitly wants bisection.
      USEDQD = (( IRANGE.EQ.ALLRNG ) .AND. (.NOT.FORCEB))

Michel Devel reports that the Intel Fortran compiler is now happy.

As a side note, we remark that this bug was fairly hard to find. USEDQD is a
LOGICAL.  Since a LOGICAL is either 1 or 0 and since both values are correct in
the xLARRE code, the code was returning correct answers in any cases.
Moreover, the code was passing most of the compilers test, which means that
most compilers are not able to check for unaffected LOGICAL.

Julien