Remove one line in the comment
From Sven

"*  Strictly speaking, Fortran does not allow aliasing of function
 *  arguments."

is not correct.  If both arguments are input, as they are in this case, the it is perfectly legal to have both actual arguments the same.  It is only if one, or both, is output that they must not be the same.

So, I suggest that we remove this sentence.