Explicit tests for a non-negative diagonal.
Appending an entry to RESULT in the testing routines was easier
than carefully shifting the placement of all the testers, but
it still required some surgery.  All entries in RESULT now are
cleared before tests are run.  All RESULT entries are checked
against the threshold, but only tests actually run are counted
in NT.

Signed-off-by: Jason Riedy <ejr@cs.berkeley.edu>