Check for trailing zeros when applying multiple Householders at once.
The idea of checking trailing zeros per Householder reflection is
extended in a simple fashion to the partitioned, bulk reflections.
Only the initial non-empty rows of the bulk reflection are used.
A particularly greedy implementation could try shrinking the order
of a bulk reflection by dropping internal TAU==0 cases, but that
involves much work for very little payoff.

Scanning the matrix to which the reflection is applied limits the
amount of work performed on band-like matrices.  The trailing zero
check reduces the work in QR of a band matrix stored densely from
O(n^3) to O(n^2*b).

Signed-off-by: Jason Riedy <ejr@cs.berkeley.edu>