Two modifications to the header of DBDSQR:
[1]

The header was reading as:

*  WORK    (workspace) DOUBLE PRECISION array, dimension (2*N)
*          if NCVT = NRU = NCC = 0, (max(1, 4*N)) otherwise

If you call this routine with NCVT = NRU = NCC = 0 and WORK of dimension (2*N)
(as indicated in the header) then it will crash on you. Explanation: If NCVT =
NRU = NCC = 0 then DBDSQR call directly DLASQ1.  The header of DLASQ1 requires
WORK to be of dimension (4*N).  So in the case NCVT = NRU = NCC = 0, you need
WORK to be of dimension (4*N).

The header is now reading as:

*  WORK    (workspace) DOUBLE PRECISION array, dimension (4*N)

[2]

If NCVT = NRU = NCC = 0, INFO is directly inherited from DLASQ1.
Therefore the description does not match.

*
*  INFO    (output) INTEGER
*          = 0:  successful exit
*          < 0:  If INFO = -i, the i-th argument had an illegal value
*          > 0:
*             if NCVT = NRU = NCC = 0,
*                = 1, a split was marked by a positive value in E
*                = 2, current block of Z not diagonalized after 30*N
*                     iterations (in inner while loop)
*                = 3, termination criterion of outer while loop not met 
*                     (program created more than N unreduced blocks)
*             else NCVT = NRU = NCC = 0,
*                   the algorithm did not converge; D and E contain the
*                   elements of a bidiagonal matrix which is orthogonally
*                   similar to the input matrix B;  if INFO = i, i
*                   elements of E have not converged to zero.
*