Julie has spotted the problem with the Intel Fortran compiler on March 2007. We also received a complain from a user on Januray 2008 (Vittorio Zecca).
In ZCGESV (line 199), the routine ZLANGE is called with sixth actual argument
WORK which is an array of COMPLEX*16.  However in ZLANGE the corresponding
argument is DOUBLE PRECISION array argument.  This is against the Fortran 95
Standard.

The interface of ZCGESV has been changed (it was a prototype driver routine)
and we request a RWORK array of DOUBLE PRECISION of dimension (N).  ZLANGE is
now called as the standard wants.