Correction reported by Elena (Oracle)
LAPACK 3.5.0 [cz]hgeqz.f compute eigenvalues differently depending on whether 
eigenvectors are requested or not. Since the computation is different, eigenvalues 
are not bit-by-bit identical whether eigenvectors are requested or not. 

This fix enables to have [cz]hgeqz.f behaving more of the same whether eigenvectors 
are requested or not. 

Note: The fact that eigenvalues are bit-by-bit identical whether eigenvectors are 
requested or not is checked in the LAPACK test suite.