address bug comment from Martin Köhler ( Max Planck Institute for Dynamics of Complex Technical Systems ) sent to lapack mailing list on Nov 29th at 7:26pm: "Bug in xGETC2" email of Martin Köhler below
(1) add a quick return for N=0

(2) add a quick return for N=1

(3) no one reviewed this fix, feel free to comment

Dear LAPACK people, 

I found a bug concerning the following function:

The xGETC2 is the level-2 BLAS implementation of an LU decomposition 
with complete pivoting. Thereby, a "comparison depending on an 
uninitialized value" appears if the input is an one-by-one matrix. 
Namely the code in line 202 (LAPACK 3.5.0): 
    IF( ABS( A( N, N ) ).LT.SMIN ) THEN 
accesses SMIN which is set in the previous DO loop. This loop is 
only called if N > 1 and so SMIN is not set to a well defined value 
if N = 1. But the documentation claims that all matrix N>=0 are 
allowed. Furthermore if calling with N = 0 the quick return 
statement is missing and the evaluation of ABS(A(N,N)) then yields
an out-of-bound memory violation. 

These bug affect the real and the double precision code as well 
as their complex counterparts from LAPACK 3.0 up to the current
LAPACK 3.6. 

regards,
Martin Köhler