oops . . .
this is related to commit r1654. 

In commit r1654, I fixed the error parameter in GESVDX to be 17 when LDVT is
invalid, I should also have changed the test suite.  Both the code and the test
suite were incorrect.

Between r1654 and r1659, LAPACK error testing were returning 4 errors because
of this.

Fixed.

Julien.