Committing Sébastien Villemot patches (5) sent on Jan 6th 2016 to LAPACK Mailing list
From Sébastien Villemot:
I attach to this message 5 patches that are currently applied to the
Debian package for LAPACK, and that I think could be usefully merged in
the LAPACK SVN tree.

The patches include fixes for old outstanding bugs in the CBLAS test
programs, fixes for the build system, and fixes for various typos.

Each patch file contains a more detailed description of its purpose.