Committing Christoph Conrads patch sent on Jan 11th on LAPACK Mailing List
From Christoph:
after the changes of the Doxygen config file in SVN revisions 1625
and 1626, routines are grouped together in the documentation, e.g., all
eigenvalue solvers for symmetric matrices in double precision.
Currently, the documentation for all routines in one group is put on
the same HTML page. I find this confusing and it slows my browser.

If every function has its own HTML page, then you can jump to the top
of the page by pressing POS1 to see the function arguments and their
description. Also you can jump to the bottom of the page with END to
see the callgraph of the respective function.

The patch in the attachment enables the SEPARATE_MEMBER_PAGES option
that puts each function documentation on its own HTML page. Please
consider applying this patch.