COMP: Prevent insource builds and inbuild installs
With cmake is best to require out-of-source builds,
and to avoid intalling into the same directory as the
build was performed.

Both of these situations can cause very confusing situations
that can frustrate new users of the software package, so
check for and provide guidance to the end-users.

These files were initially contributed to the ITK
(www.itk.org) project and have been widely re-used
by many other projects.

NOTE: This patch is dependant on pull request #16