bug: fixes always failing tests for *TREXC which was erroneous tests
There where checks for *TREXC for the argument of N = 0 with a return value
of -7. However, the documentation of the *TREXC routines specify that
N = 0 is a valid argument with instant return.

Hence the checks have been removed and the quick returns established in the
affected routines.

Also comments for [dz]errec.f files are fixed.

Signed-off-by: Nick Papior <nickpapior@gmail.com>