Fixes #92: xLASWP: bug for negative INCX
Thanks to Vladimír Chalupecký for noticing quite an old bug in xLASWP.  The bug
is fixed now. The cases for INCX = 1 and INCX = -1 were correct.  It is not
clear whether any other cases would be useful. So the bug was never detected.
We might want to consider to (1) forbid any values for INCX but INCX = 1 and
INCX = -1, (2) and simplify the code and the subroutine description
accordingly. Comments welcome. Julien.