Allow building LAPACKE with XBLAS and TMGLIB simultaneously
Order the libraries the same as the Makefiles are ordering them:
tmglib, lapack, xblas, blas.

Fix combinations of USE_XBLAS with BUILD_SINGLE, etc. so that it only
builds the LAPACK extended functions of the desired type and precision.

Remove "OBJ" from variable names that are not lists of objects.