Use appending to create lists of source files
Allows any combination of types/precisions to be built at once.

Name the lists "SOURCES" instead of "ALLOBJ" since they contain lists of
source files, not object files.

Fix problems in BLAS:  Was missing the ${CBLAS3} file set when building
with BLAS_COMPLEX.  Was adding ${BLASLIB} as if it were a source file.