Fix #52
Added in comment of xLARRV

Note: VU is currently not used by this implmentation of DLARRV, VU is
passed to DLARRV because it could be used compute gaps on the right end
of the extremal eigenvalues. However, with not much initial accuracy in
LAMBDA and VU, the formula can lead to an overestimation of the right gap
and thus to inadequately early RQI 'convergence'. This is currently
prevented this by forcing a small right gap. And so it turns out that VU
is currently not used by this implementation of DLARRV.