Fix #147: xlapy2 not propagating nans
xLAPY2 now returns a NaN whenever input variables X or Y are NaNs.

The previous xLAPY2 was relying on FORTRAN INTRINSIC MAX and MIN to behave in a
certain way with NaNs (i.e. return a NaN whenever X or Y are NaN) and this
behavior is not observed on some (most?) compilers.

We handle the NaN behavior of xLAPY2 by checking for NaNs at the start of the
function.

Thanks to Andreas Noack for providing report and sample code to demonstrate the
problem.