Follow-up to #367; fix lapack.h for deprecated functions [cdsz]ggsv[dp]
These functions were inadvertently removed before 3.9.0 and readded
by #367. However, some of the LAPACK-functions in lapack.h were
declared as LAPACKE-functions (which also have a different signature
than their LAPACK-counterparts), which leads (at least) to the compiler
emitting a [-Wimplicit-function-declaration] warning and not knowing
which function to choose:

[...]/LAPACKE/src/lapacke_cggsvp_work.c: In function 'LAPACKE_cggsvp_work':
[...]/LAPACKE/include/lapack.h:3761:37: warning: implicit declaration of function 'cggsvp_'; did you mean 'cggsvp3_'? [-Wimplicit-function-declaration]
 #define LAPACK_cggsvp LAPACK_GLOBAL(cggsvp,CGGSVP)
                                     ^
[...]/LAPACKE/include/lapacke_mangling.h:12:39: note: in definition of macro 'LAPACK_GLOBAL'
 #define LAPACK_GLOBAL(lcname,UCNAME)  lcname##_
                                       ^~~~~~
[...]/LAPACKE/src/lapacke_cggsvp_work.c:52:9: note: in expansion of macro 'LAPACK_cggsvp'
         LAPACK_cggsvp( &jobu, &jobv, &jobq, &m, &p, &n, a, &lda, b, &ldb, &tola,
         ^~~~~~~~~~~~~

Here, we align the declarations in lapack.h with their callsites in
the various LAPACKE_[cdsz]ggsv[dp]_work functions again.

For more discussion see
https://github.com/conda-forge/lapack-feedstock/pull/32

Suggested-By: Isuru Fernando <isuruf@gmail.com>