Updated values of TWOPI and PIOVER2 to 39 significant digits, to prevent accuracy loss when REAL or DOUBLE PRECISION types correspond to an extended precision real kind. These constants should be sufficient to work with 128-bit extended-IEEE data type.
Rebuilding the modified source using the default gfortran make.inc
(with gfortran 7.5.0) does not add any new test failures. The
only test failure, which is also present in the master baranch, is:

 ZDRGEV3: ZGGEV31 returned INFO=     7.
   N=    12, JTYPE=    16, ISEED=( 2855, 3551, 2406, 2393)

Additionally, included a sample build file for quadruple-precision
compilation using gfortran (version 4.7 or later). With this
config file, both REAL and DOUBLE PRECISION are mapped onto
a 128-bit REAL(16) type, with the corresponding change to the
COMPLEX and COMPLEX*16.

There are numerical failure 6 failures for each type, all in
*GBSVX type 6 matrix, with S/D routines losing 4 significant
digits, and C/Z routines 5 digits relative to the expected
result. I am not really qualified to debug this failure ...