cmake build: more modularity, linker fix, pkg-config fixup
This set of changes stems from the packaging of netlib code in pkgsrc,
where we decided to provide blas, lapack, cblas, and lapacke as separate
packages and hence need clean support in the build system for this and
also for combining the wrappers lapacke and cblas with any prescribed
optimized blas or lapack library.

- LAPACK option to be able to just build BLAS
- consistent logic for USE_OPTIMIZED_BLAS and USE_OPTIMIZED_LAPACK
  (including clear failure instead of silent fallback if BLAS_LIBARIES
  or LAPACK_LIBRARIES do not work)
- Fix linking of CBLAS on NetBSD with pkgsrc, use Fortran and not C
  for linking (same as LAPACKE already).
- Allow building of shared and static libs at the same time.
- Trying to fix up pkg-config files for all cases of internal or
  external libs

The last one is not in really proper shape yet. The idea is that the
.pc files contain

  Libs.private= -L/fooprefix -lfoo

if foo is your optimized BLAS/LAPACK, not assuming presence of a .pc file
for that itself. This works for user-supplied BLAS_LIBRARIES so far, but
inserts the full path to the .so file from its search … which does not fit
the case of static linking. And well, yes, you do not know if both static
and shared libs are present. So maybe it's a complex problem. My use case
so far is fine as I explicitly provide BLAS_LIBRARIES and actually this is
what I try to get into other upstream packages that need BLAS or LAPACK:
Just use BLAS_LIBS or LAPACK_LIBS as provided by the user (the packager).

Would be nice if (most of) this can be included in the official LAPACK CMake
build so that we do not have to carry and update the patches all the time.