Fix CBLAS xerbla following hidden strlen argument changes
Commits 2d9fbdeecca906d7abb89e9cc16b02ed5286298e and
628a2095c24021ab07904e036389ebc4feb3e67a changed the function prototypes in the
cblas_f77.h.

However, the F77_xerbla case was not treated in the same way as the others. In
particular, the consequence is now that the xerbla function provided by the
CBLAS/src/xerbla.c file is never called at runtime (because it is identified as
“F77_xerbla” in the symbol table, instead of simply “xerbla”, since F77_xerbla
is *not* a macro).

This patch restores the symmetry between xerbla and the other functions.