Improve FMA usage in laqr5
Rearrange the application of the Householder reflector
to save one instruction per dot product if FMA is
available.

The update from the right, H * (I - tau * v * v**T),
for example, changes from
    H - (tau * (H * v)) * v**T
to
    H - (H * (v * tau)) * v**T.
The instruction savings are due to the special structure
of v, whose first component is implicitly one (and used
for storing tau).