Move single-precision isnan to ALLAUX
It was in the SCLAUX group until 09af64a8ef305d9bceb1ad69c178f89451ee6df6
when it was realized that things in both DSLASRC and ZCLASRC need it.
Therefore it's used when building for any S, D, C, or Z precision.

Using similar rationale, la_constants is in both SCLAUX and DZLAUX, and
can be moved to ALLAUX.