Fix CMake installation instructions in LAPACK README
The `CMAKE_INSTALL_LIBDIR` was used here from the command line, probably
accidentally. It is a variable provided by the core CMake module
GNUInstallDirs and shouldn't be overridden. The variable, that makes
sense in this context is the `CMAKE_INSTALL_PREFIX` variable, which is
intended to be set by the user.