fix library names in `lapack64.pc`
Add substitutions to `lapack.pc.in`, in order to ensure that
the generated `lapack64.pc` file correctly references the index64
library and BLAS dependency.  It seems that other pkg-config templates
have been updated as part of #462 but this one was omitted.
As a result, finding a lapack64 via pkg-config gave non-index64
libraries.

Signed-off-by: Michał Górny <mgorny@gentoo.org>