Implement MSVC complex types
Implement the support for the `_Fcomplex` and `_Dcomplex` types used
by MSVC in place of the standard C99 complex types.  This is necessary
to make LAPACK build out of the box with MSVC, since it does not
implement the default C99 types.

See: https://learn.microsoft.com/en-us/cpp/c-runtime-library/complex-math-support

I went for the simplest implementation possible, limiting the changes
to swapping the default types used when MSVC is used as the compiler,
also when `lapack_config.h` is not used.  I haven't added
a `LAPACK_COMPLEX_*` ifdef for it, but I can do that if you prefer.

Signed-off-by: Michał Górny <mgorny@quansight.com>